#83: refactor serializers into multiple files
#83: refactor serializers into multiple files 2
