+++ src/api/tests/serializers/test_simulation_detail.py
@@ -5,7 +5,7 @@ from django.urls import path, include, reverse
 from django.utils import timezone
 
 from api.models.models import LAWMSimulation
-from api.serializers import SimulationDetailSerializer
+from api.serializers.simulation_serializers import SimulationDetailSerializer
 from api.tests.api_test_mixin import MicroSimuTestMixin
 
 # The hyperlinked serializer depends on the urls, so we need to specify the urls
