+++ src/api/tests/tests_views.py
@@ -7,8 +7,9 @@ from rest_framework.reverse import reverse as drf_reverse
 from rest_framework.test import APITestCase
 
 from api.models.models import LAWMSimulation
-from api.serializers import SimulationListSerializer, SimulationDetailSerializer, RegionResultSerializer, \
-    RunParametersSerializer
+from api.serializers.parameters_serializers import RunParametersSerializer
+from api.serializers.results_serializers import RegionResultSerializer
+from api.serializers.simulation_serializers import SimulationListSerializer, SimulationDetailSerializer
 
 from api.std_lib.lawm.regions import Africa
 from api.tests.api_test_mixin import MicroSimuTestMixin
